+++ src/main/java/info/coremodding/craftenchants/event/initialize/PostInitializer.java
@@ -2,14 +2,13 @@
  * Copyright 2014 by CoreModding under GNU Lesser General Public License (LGPLv3)
  * http://www.gnu.org/licenses/gpl.html http://www.gnu.org/licenses/lgpl.html
  */
-package info.coremodding.craftenchants.event;
+package info.coremodding.craftenchants.event.initialize;
 
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 
 public class PostInitializer {
   public static void handle(FMLPostInitializationEvent postEvent) {
-    // TODO Implement method handle
-
+    
   }
 
 }
